+++ mrv2/lib/mrvPanels/mrvPythonPanel.h
@@ -45,6 +45,10 @@ namespace mrv
             void comment_text();
             void uncomment_text();
 
+            void jump_to_error();
+
+            void external_editor();
+
             void add_to_script_list(const std::string& file);
             void script_shortcut(unsigned int idx);
 
@@ -76,6 +80,10 @@ namespace mrv
             static void clear_editor_cb(Fl_Menu_*, PythonPanel* o);
 
             static void toggle_line_numbers_cb(Fl_Menu_*, PythonPanel* o);
+
+            static void jump_to_error_cb(Fl_Menu_*, PythonPanel* o);
+
+            static void external_editor_cb(Fl_Menu_*, PythonPanel* o);
 
             static void add_to_script_list_cb(Fl_Menu_* m, PythonPanel* p);
             static void script_shortcut_cb(Fl_Menu_* m, PythonPanel* d);