+++ mrv2/python/plug-ins/update-mrv2.py
@@ -3,7 +3,7 @@
 # Copyright Contributors to the mrv2 Project. All rights reserved.
 
 
-
+import crap
 
 #
 # Standard libs